pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ableguides"
version = "0.1.0"
description = "Generate spoken guide cue WAVs via ElevenLabs and assemble Ableton Drum Rack presets"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "mido>=1.3",
]

[project.scripts]
ableguides = "ableguides.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
target-version = "py311"
line-length = 100
